+++ src/core/release/resolver.rs
@@ -3,7 +3,7 @@ use crate::engine::pipeline::PipelineCapabilityResolver;
 use crate::error::{Error, Result};
 use crate::extension::{self, ExtensionManifest};
 
-use super::release_step_type::ReleaseStepType;
+use super::types::ReleaseStepType;
 
 pub(crate) struct ReleaseCapabilityResolver {
     extensions: Vec<ExtensionManifest>,
